+++ src/randombytes.rs
@@ -1,6 +1,5 @@
 //! Cryptographic random number generation.
 use ffi;
-use libc::size_t;
 use std::iter::repeat;
 
 /// `randombytes()` randomly generates size bytes of data.
@@ -12,7 +11,7 @@ pub fn randombytes(size: usize) -> Vec<u8> {
     unsafe {
         let mut buf: Vec<u8> = repeat(0u8).take(size).collect();
         let pbuf = buf.as_mut_ptr();
-        ffi::randombytes_buf(pbuf, size as size_t);
+        ffi::randombytes_buf(pbuf, size);
         buf
     }
 }
@@ -24,6 +23,6 @@ pub fn randombytes(size: usize) -> Vec<u8> {
 /// from sodiumoxide.
 pub fn randombytes_into(buf: &mut [u8]) {
     unsafe {
-        ffi::randombytes_buf(buf.as_mut_ptr(), buf.len() as size_t);
+        ffi::randombytes_buf(buf.as_mut_ptr(), buf.len());
     }
 }
